requirements.txt
numpy==1.26.2
scipy==1.11.4
matplotlib
pandas
sentence_transformers==2.5.1
langchain_core==0.1.32
langchain_mistralai==0.0.5
anthropic==0.20.0
google-generativeai==0.4.1
seaborn